requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch==2.0.1
torchvision==0.15.2
albumentations==0.4.3
opencv-python
pudb==2019.2
imageio==2.9.0
imageio-ffmpeg==0.4.2
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
torch-fidelity==0.3.0
transformers==4.22.2
kornia==0.8.1
webdataset==0.2.5
torchmetrics==1.7.2
fire==0.4.0
gradio==3.21.0
diffusers==0.12.1
datasets[vision]==2.4.0
carvekit-colab==4.1.0
rich>=13.3.2
lovely-numpy>=0.2.8
lovely-tensors>=0.1.14
plotly==5.13.1
ninja
accelerate
ipdb
tqdm
wandb
open-clip-torch
taming-transformers-rom1504
git+https://github.com/openai/CLIP.git
trimesh
icecream
pyglet
pyrender
pycolmap
enlighten
open3d
poselib
lpips
seaborn